+++ src/main/resources/templates/index.html
@@ -55,9 +55,9 @@
     <section class="py-5">
         <div class="container">
             <!-- 섹션 제목과 More 버튼 배치 -->
-            <div class="d-flex justify-content-between align-items-center mb-5">
-                <h2 class="text-center flex-grow-1">추천 객실</h2>
-                <a href="/roomList" class="btn btn-outline-primary ms-auto">More</a>
+            <div class="d-flex justify-content-between align-items-center mb-5 recommended-wrap">
+                <h2 class="text-center flex-grow-1">Recommended room</h2>
+                <a href="/roomList" class="btn btn-outline-primary ms-auto more-btn">More</a>
             </div>
 
             <!-- 추천 방이 있을 경우 -->
